+++ frontend/openchat-client/src/utils/chat.ts
@@ -1030,13 +1030,6 @@ export function mergeSendMessageResponse(
     };
 }
 
-export function markAllRead(chat: ChatSummary): void {
-    const latestMessageIndex = chat.latestMessage?.event.messageIndex;
-    if (latestMessageIndex !== undefined) {
-        messagesRead.markReadUpTo(chat.chatId, latestMessageIndex);
-    }
-}
-
 export function mergeEventsAndLocalUpdates(
     events: EventWrapper<ChatEvent>[],
     unconfirmed: EventWrapper<Message>[],